+++ testutil/keeper/crosschain.go
@@ -396,7 +396,6 @@ func MockSaveOutbound(
 	m.On("RemoveFromPendingNonces",
 		ctx, tss.TssPubkey, cctx.GetCurrentOutboundParam().ReceiverChainId, mock.Anything).
 		Return().Times(expectedNumberOfOutboundParams)
-	//m.On("GetTSS", ctx).Return(observertypes.TSS{}, true)
 }
 
 func MockSaveOutboundNewRevertCreated(