+++ x/crosschain/genesis.go
@@ -44,13 +44,12 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	// Set all the cross-chain txs
 	for _, elem := range genState.CrossChainTxs {
 		if elem != nil {
-			k.SetCctxAndNonceToCctxAndInboundHashToCctx(ctx, *elem, func(ctx sdk.Context) string {
-				tss, found := k.GetObserverKeeper().GetTSS(ctx)
-				if !found {
-					return ""
-				}
-				return tss.TssPubkey
-			})
+			tss, found := k.GetObserverKeeper().GetTSS(ctx)
+			if found {
+				k.SetCctxAndNonceToCctxAndInboundHashToCctx(ctx, *elem, func(ctx sdk.Context) string {
+					return tss.TssPubkey
+				})
+			}
 		}
 	}
 	for _, elem := range genState.FinalizedInbounds {