+++ x/crosschain/keeper/grpc_query_inbound_hash_to_cctx_test.go
@@ -126,7 +126,11 @@ func TestInTxHashToCctxQueryPaginated(t *testing.T) {
 	})
 }
 
-func createInTxHashToCctxWithCctxs(ctx sdk.Context, keeper *crosschainkeeper.Keeper, tssPubkey string) ([]types.CrossChainTx,
+func createInTxHashToCctxWithCctxs(
+	ctx sdk.Context,
+	keeper *crosschainkeeper.Keeper,
+	tssPubkey string,
+) ([]types.CrossChainTx,
 	types.InboundHashToCctx) {
 	cctxs := make([]types.CrossChainTx, 5)
 	for i := range cctxs {