+++ changelog.md
@@ -10,15 +10,17 @@
 * [2634](https://github.com/zeta-chain/node/pull/2634) - add support for EIP-1559 gas fees
 * [2597](https://github.com/zeta-chain/node/pull/2597) - Add generic rpc metrics to zetaclient
 * [2538](https://github.com/zeta-chain/node/pull/2538) - add background worker routines to shutdown zetaclientd when needed for tss migration
-* [2644](https://github.com/zeta-chain/node/pull/2644) - add created_timestamp to cctx status
+
 
 ### Refactor
 
 * [2615](https://github.com/zeta-chain/node/pull/2615) - Refactor cleanup of outbound trackers
+* [2644](https://github.com/zeta-chain/node/pull/2644) - refactor SetCctxAndNonceToCctxAndInboundHashToCctx receive tsspubkey as an argument
 
 ### Fixes
 
 * [2654](https://github.com/zeta-chain/node/pull/2654) - add validation for authorization list in when validating genesis state for authorization module
+* [2725](https://github.com/zeta-chain/node/pull/2725) - 
 
 ## v19.0.0